fixc & separate test for addresing modes
